refactor: change dev4 to ui
