+++ packages/devkit/package.json
@@ -42,10 +42,7 @@
     "execa": "^8.0.1",
     "fast-glob": "^3.3.3",
     "fs-extra": "^11.3.0",
-    "ink": "^5.2.1",
-    "ink-text-input": "^5.0.1",
     "lodash": "^4.17.21",
-    "neo-blessed": "^0.2.0",
     "npm-packlist": "^10.0.1",
     "open": "^10.2.0",
     "p-all": "^3.0.0",
