+++ packages/devkit/tsconfig.json
@@ -9,7 +9,6 @@
     "declaration": true,
     "noEmit": true,
     "esModuleInterop": true,
-    "jsx": "react",
     "skipLibCheck": true
   },
   "exclude": ["./dist"],
